third_party/blink/web_tests/fast/dom/shadow/shadow-root-new-expected.txt

CONSOLE WARNING: line 19: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test checks the object 'host.createShadowRoot()' and internals.shadowRoot(div) are the same.

PASS root1 is root2
PASS successfullyParsed is true

TEST COMPLETE
